requirements.txt
requests
newspaper3k
pika
schedule
lxml_html_clean
